requirements.txt
Django==4.2.1
keras_vggface==0.6
matplotlib==3.7.0
mtcnn==0.1.1
numpy==1.21.5
opencv_python==4.7.0.72
Pillow==9.5.0
requests==2.28.1
tensorflow==2.10.0
tensorflow_datasets==4.9.2
torch==1.13.1
torchvision==0.14.1
